Move crispy code into id engine files.
This commit refactors the crispy SDL2 code into regular id engine
files. All of the game code uses the engine files to interact with the
platform, so there is no point to having a seperate crispy abstraction
layer. Mouse wheel weapon change is also implemented as part of this
commit.